+++ src/pytorch_lightning/strategies/fully_sharded_native.py
@@ -17,12 +17,10 @@ from typing import Any, Dict, Generator, List, Optional, Union
 
 import torch
 from torch import Tensor
-from torch.distributed.distributed_c10d import _get_default_group, ProcessGroup
 
 import pytorch_lightning as pl
 from lightning_lite.utilities.distributed import (
     _get_process_group_backend_from_env,
-    distributed_available,
     get_default_process_group_backend_for_device,
 )
 from lightning_lite.utilities.distributed import group as _group
@@ -42,9 +40,11 @@ from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from pytorch_lightning.utilities.imports import _TORCH_GREATER_EQUAL_1_12
 from pytorch_lightning.utilities.model_helpers import is_overridden
 from pytorch_lightning.utilities.rank_zero import rank_zero_info, rank_zero_only
-from pytorch_lightning.utilities.types import STEP_OUTPUT
+from pytorch_lightning.utilities.types import ProcessGroup, STEP_OUTPUT
 
-if _TORCH_GREATER_EQUAL_1_12:
+_distributed_available = torch.distributed.is_available()
+_fsdp_available = _TORCH_GREATER_EQUAL_1_12 and _distributed_available
+if _fsdp_available:
     from torch.distributed.fsdp.fully_sharded_data_parallel import (
         BackwardPrefetch,
         CPUOffload,
@@ -57,6 +57,9 @@ else:
     MixedPrecision = None  # type: ignore[misc,assignment]
     BackwardPrefetch = None  # type: ignore[misc,assignment]
     CPUOffload = None  # type: ignore[misc,assignment]
+
+if _distributed_available:
+    from torch.distributed.distributed_c10d import _get_default_group
 
 log = logging.getLogger(__name__)
 
@@ -274,7 +277,7 @@ class DDPFullyShardedNativeStrategy(ParallelStrategy):
             yield
 
     def barrier(self, name: Optional[str] = None) -> None:
-        if not distributed_available():
+        if not _distributed_available:
             return
         if torch.distributed.get_backend() == "nccl":
             torch.distributed.barrier(device_ids=self._determine_device_ids())
@@ -357,7 +360,7 @@ class DDPFullyShardedNativeStrategy(ParallelStrategy):
 
     @classmethod
     def register_strategies(cls, strategy_registry: Dict) -> None:
-        if _TORCH_GREATER_EQUAL_1_12:
+        if _fsdp_available:
             strategy_registry.register(
                 "fsdp_native",
                 cls,