+++ src/pytorch_lightning/utilities/types.py
@@ -23,12 +23,16 @@ from typing import Any, Dict, Generator, List, Mapping, Optional, Sequence, Type
 
 import torch
 from torch import Tensor
-from torch._C._distributed_c10d import ProcessGroup
 from torch.utils.data import DataLoader
 from torchmetrics import Metric
 from typing_extensions import Protocol, runtime_checkable
 
 from lightning_lite.utilities.types import _LRScheduler, ReduceLROnPlateau
+
+if torch.distributed.is_available():
+    from torch._C._distributed_c10d import ProcessGroup
+else:
+    ProcessGroup = ...  # type: ignore[assignment,misc]
 
 _NUMBER = Union[int, float]
 _METRIC = Union[Metric, Tensor, _NUMBER]